Added google maps components
